// src/app/services/API/models/get-market-dto.ts
/* tslint:disable */
export interface GetMarketDto {
    apiActive?: number;
    bourseCode?: string;
    brokerId: number;
    brokerName: string;
    id: number;
    isBOC?: number;
    nationalId: string;
    organizationType: string;
    pamCode: string;
    password?: string;
    symbolORFundTitle?: string;
    username?: string;
}